requirements.txt
hdfs
elasticsearch
beautifulsoup4
kafka-python